// apps/react-frontend/src/api/client/models/authorization-url-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Eazy Recycling Backend Service
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface AuthorizationUrlResponse
 */
export interface AuthorizationUrlResponse {
    /**
     * 
     * @type {string}
     * @memberof AuthorizationUrlResponse
     */
    'authorizationUrl': string;
    /**
     * 
     * @type {string}
     * @memberof AuthorizationUrlResponse
     */
    'state': string;
}